Horace access Python class with Matlab wrapper
This implements an abstract Matlab class called light_python_wrapper
which is meant to be used to access any Python class.
To use it to wrap a class, subclass it and define a constructor
which points to the desired Python class.
An example is in euphonic.m and ureg.m which wraps euphonic_wrapper.py
and euphonic.ureg respectively.
euphonic_wrapper is a Python class which wraps the ForceConstants
and QpointPhononModes classes and provides a Horace access function
horace_sqw