Make is_initialised property not Constant
For some reason if the property is Constant, it is evaluated multiple
times if an error is raised during execution. If you emit any kind of
output (e.g. a warning) before it gets to the error it will be emitted
multiple (specifically 3) times. Don't know why this is happening
exactly but given it is controlled by a global so has small overhead
just make the properties not constant.

In this case I triggered this in euphonic_on by having one module
present but the incorrect version (meaning verify_python_modules
emitted a warning), and another module that wasn't present at all
(meaning verify_python_modules raised an error). The warning was
inexplicably emitted 3 times before the error was raised.